+++ packages/constellation/src/components/Base/Icon/svg.d.ts
@@ -1,4 +1,4 @@
 declare module '*.svg' {
-  const content: any
+  const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>
   export default content
 }
